Move PerReplicable to replication package
