+++ slingjunit.parent/core/src/main/java/com/peregrine/slingjunit/localFS/LocalFSJTest.java
@@ -1,6 +1,6 @@
 package com.peregrine.slingjunit.localFS;
 
-import com.peregrine.adaption.PerReplicable;
+import com.peregrine.replication.PerReplicable;
 import com.peregrine.replication.Replication;
 import com.peregrine.replication.ReplicationsContainer;
 import com.peregrine.slingjunit.ReplicationTestBase;
